Add DMA code support
There are a couple other fields on the object, but they are fairly uncommon (like metro codes, which seem to be a slightly different and substantially less common form of dma code). I added this one as it's necessary for my work. Other attributes can be found in the code here:

https://github.com/maxmind/geoip-api-java/blob/master/src/main/java/com/maxmind/geoip/Location.java